+++ jdplus-main-base/jdplus-spreadsheet-base-parent/jdplus-spreadsheet-base-api/src/main/java/module-info.java
@@ -1,4 +1,6 @@
+import jdplus.sa.base.api.SaOutputFactory;
 import jdplus.spreadsheet.base.api.SpreadSheetProvider;
+import jdplus.spreadsheet.base.api.sa.SpreadsheetOutputFactory;
 import jdplus.toolkit.base.api.timeseries.TsProvider;
 
 module jdplus.spreadsheet.base.api {
@@ -9,10 +11,15 @@ module jdplus.spreadsheet.base.api {
     requires static org.checkerframework.checker.qual;
 
     requires jdplus.toolkit.base.tsp;
+    requires jdplus.sa.base.core;
     requires nbbrd.spreadsheet.api;
     requires nbbrd.io.base;
 
     exports jdplus.spreadsheet.base.api;
+    exports jdplus.spreadsheet.base.api.sa;
+
+    provides SaOutputFactory
+            with SpreadsheetOutputFactory;
 
     provides TsProvider
             with SpreadSheetProvider;