+++ jdplus-main-base/jdplus-toolkit-base-parent/jdplus-toolkit-base-api/src/main/java/jdplus/toolkit/base/api/timeseries/TsDataTable.java
@@ -16,12 +16,14 @@
  */
 package jdplus.toolkit.base.api.timeseries;
 
+import jdplus.toolkit.base.api.math.matrices.Matrix;
 import jdplus.toolkit.base.api.util.Collections2;
-import jdplus.toolkit.base.api.util.List2;
 import jdplus.toolkit.base.api.util.function.BiIntPredicate;
 import lombok.AccessLevel;
+import lombok.NonNull;
+import nbbrd.design.StaticFactoryMethod;
+import nbbrd.design.VisibleForTesting;
 import org.checkerframework.checker.index.qual.NonNegative;
-import org.checkerframework.checker.nullness.qual.NonNull;
 
 import java.time.LocalDateTime;
 import java.util.Iterator;
@@ -31,11 +33,11 @@ import java.util.function.Function;
 import java.util.function.IntFunction;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
-import jdplus.toolkit.base.api.math.matrices.Matrix;
 
 /**
  * @author Philippe Charles
  */
+@lombok.Getter
 @lombok.AllArgsConstructor(access = AccessLevel.PRIVATE)
 public final class TsDataTable {
 
@@ -47,33 +49,27 @@ public final class TsDataTable {
         PRESENT, UNUSED, BEFORE, AFTER, EMPTY
     }
 
-    @NonNull
-    public static <X> TsDataTable of(@NonNull Iterable<X> col, @NonNull Function<? super X, TsData> toData) {
-        TsDomain domain = computeDomain(Collections2.streamOf(col).map(toData).map(TsData::getDomain).filter(o -> !o.isEmpty()).iterator());
-        return new TsDataTable(domain, Collections2.streamOf(col).map(toData).collect(List2.toUnmodifiableList()));
+    @StaticFactoryMethod
+    public static @NonNull <X> TsDataTable of(@NonNull Iterable<X> col, @NonNull Function<? super X, TsData> toData) {
+        TsDomain domain = computeDomain(Collections2.streamOf(col).map(toData).filter(Objects::nonNull).map(TsData::getDomain).filter(o -> !o.isEmpty()).iterator());
+        return new TsDataTable(domain, Collections2.streamOf(col).map(toData).toList());
     }
 
-    @NonNull
-    public static TsDataTable of(@NonNull Iterable<TsData> col) {
+    @StaticFactoryMethod
+    public static @NonNull TsDataTable of(@NonNull Iterable<TsData> col) {
         return of(col, Function.identity());
     }
 
-    @lombok.NonNull
-    @lombok.Getter
-    private final TsDomain domain;
+    private final @NonNull TsDomain domain;
 
-    @lombok.NonNull
-    @lombok.Getter
-    private final List<TsData> data;
+    private final @NonNull List<TsData> data;
 
-    @NonNull
-    public Cursor cursor(@NonNull DistributionType distribution) {
+    public @NonNull Cursor cursor(@NonNull DistributionType distribution) {
         Objects.requireNonNull(distribution);
         return cursor(i -> distribution);
     }
 
-    @NonNull
-    public Cursor cursor(@NonNull IntFunction<DistributionType> distribution) {
+    public @NonNull Cursor cursor(@NonNull IntFunction<DistributionType> distribution) {
         Objects.requireNonNull(distribution);
         return new Cursor(getDistributors(data, distribution));
     }
@@ -165,18 +161,14 @@ public final class TsDataTable {
     }
 
     private static BiIntPredicate getDistributor(DistributionType type) {
-        switch (type) {
-            case FIRST:
-                return (pos, size) -> pos % size == 0;
-            case LAST:
-                return (pos, size) -> pos % size == size - 1;
-            case MIDDLE:
-                return (pos, size) -> pos % size == size / 2;
-            default:
-                throw new RuntimeException();
-        }
+        return switch (type) {
+            case FIRST -> (pos, size) -> pos % size == 0;
+            case LAST -> (pos, size) -> pos % size == size - 1;
+            case MIDDLE -> (pos, size) -> pos % size == size / 2;
+        };
     }
 
+    @VisibleForTesting
     static TsDomain computeDomain(Iterator<TsDomain> domains) {
         if (!domains.hasNext()) {
             return TsDomain.DEFAULT_EMPTY;
@@ -254,5 +246,4 @@ public final class TsDataTable {
         }
         return Matrix.of(m, nr, nc);
     }
-
 }