+++ CHANGELOG.md
@@ -17,12 +17,13 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - ![STAT] Correct deviances in diffuse likelihood
 - ![STAT] Correct covariance of ARMA parameters in models with quasi-unit roots in AR
 - ![UI] Save correctly modified multi-processing
+- ![IO] Fix NPE in grid reader when header is null
 
 ### Added
 
 - ![STAT] Add Poisson distribution
 - ![OTHER] Serialize High-frequency series (modelling part)
-
+- ![UI] Add export of SA documents to Excel
 
 ## [3.0.2] - 2023-06-14